refactor: Split up admin dashboard into several files for easier development
